Rename package 'geometric' to 'geom'.
