+++ src/main/java/net/imagej/ops/AbstractOpEnvironment.java
@@ -44,7 +44,7 @@ import net.imagej.ops.features.haralick.HaralickNamespace;
 import net.imagej.ops.features.zernike.ZernikeNamespace;
 import net.imagej.ops.features.tamura2d.TamuraNamespace;
 import net.imagej.ops.filter.FilterNamespace;
-import net.imagej.ops.geometric.Geometric2DNamespace;
+import net.imagej.ops.geom.Geometric2DNamespace;
 import net.imagej.ops.geometric3d.Geometric3DNamespace;
 import net.imagej.ops.image.ImageNamespace;
 import net.imagej.ops.imagemoments.ImageMomentsNamespace;
