+++ src/main/java/net/imagej/ops/geom/DefaultCircularity.java
@@ -27,12 +27,12 @@
  * POSSIBILITY OF SUCH DAMAGE.
  * #L%
  */
-package net.imagej.ops.geometric;
+package net.imagej.ops.geom;
 
 import net.imagej.ops.FunctionOp;
 import net.imagej.ops.Ops.Geometric2D;
-import net.imagej.ops.Ops.Geometric2D.ConvexHull;
-import net.imagej.ops.Ops.Geometric2D.Convexity;
+import net.imagej.ops.Ops.Geometric2D.Area;
+import net.imagej.ops.Ops.Geometric2D.Circularity;
 import net.imagej.ops.Ops.Geometric2D.Perimeter;
 import net.imagej.ops.RTs;
 import net.imglib2.roi.geometric.Polygon;
@@ -41,34 +41,29 @@ import net.imglib2.type.numeric.RealType;
 import org.scijava.plugin.Plugin;
 
 /**
- * Generic implementation of {@link Convexity}.
+ * Generic implementation of {@link Circularity}.
  * 
  * @author Daniel Seebacher, University of Konstanz.
  */
-@Plugin(type = GeometricOp.class, label = "Geometric: Convexity", name = Geometric2D.Convexity.NAME)
-public class DefaultConvexity<O extends RealType<O>> extends
-		AbstractGeometricFeature<Polygon, O> implements Geometric2D.Convexity {
+@Plugin(type = GeometricOp.class, label = "Geometric: Circularity", name = Geometric2D.Circularity.NAME)
+public class DefaultCircularity<O extends RealType<O>> extends
+		AbstractGeometricFeature<Polygon, O> implements Geometric2D.Circularity {
 
-	private FunctionOp<Polygon, Polygon> convexHullFunction;
+	private FunctionOp<Polygon, O> areaFunc;
 	private FunctionOp<Polygon, O> perimiterFunc;
 
 	@Override
 	public void initialize() {
-		convexHullFunction = ops().function(ConvexHull.class, Polygon.class,
-				Polygon.class);
+		areaFunc = RTs.function(ops(), Area.class, in());
 		perimiterFunc = RTs.function(ops(), Perimeter.class, in());
 	}
 
 	@Override
 	public void compute(final Polygon input, final O output) {
-
-		// get perimeter of input and its convex hull
-		final O inputArea = perimiterFunc.compute(input);
-		final O convexHullArea = perimiterFunc.compute(convexHullFunction
-				.compute(input));
-
-		output.setReal(convexHullArea.getRealDouble()
-				/ inputArea.getRealDouble());
+		output.setReal(4
+				* Math.PI
+				* (areaFunc.compute(input).getRealDouble() / Math.pow(
+						perimiterFunc.compute(input).getRealDouble(), 2)));
 	}
 
 }
