+++ src/main/java/net/imagej/ops/geom/DefaultConvexHull.java
@@ -27,7 +27,7 @@
  * POSSIBILITY OF SUCH DAMAGE.
  * #L%
  */
-package net.imagej.ops.geometric;
+package net.imagej.ops.geom;
 
 import java.util.ArrayList;
 import java.util.Collections;
