+++ src/main/java/net/imagej/ops/geom/DefaultElongation.java
@@ -27,13 +27,13 @@
  * POSSIBILITY OF SUCH DAMAGE.
  * #L%
  */
-package net.imagej.ops.geometric;
+package net.imagej.ops.geom;
 
 import net.imagej.ops.FunctionOp;
 import net.imagej.ops.Ops.Geometric2D;
-import net.imagej.ops.Ops.Geometric2D.Area;
-import net.imagej.ops.Ops.Geometric2D.Circularity;
-import net.imagej.ops.Ops.Geometric2D.Perimeter;
+import net.imagej.ops.Ops.Geometric2D.Elongation;
+import net.imagej.ops.Ops.Geometric2D.MajorAxis;
+import net.imagej.ops.Ops.Geometric2D.MinorAxis;
 import net.imagej.ops.RTs;
 import net.imglib2.roi.geometric.Polygon;
 import net.imglib2.type.numeric.RealType;
@@ -41,29 +41,29 @@ import net.imglib2.type.numeric.RealType;
 import org.scijava.plugin.Plugin;
 
 /**
- * Generic implementation of {@link Circularity}.
+ * Generic implementation of {@link Elongation}.
  * 
  * @author Daniel Seebacher, University of Konstanz.
  */
-@Plugin(type = GeometricOp.class, label = "Geometric: Circularity", name = Geometric2D.Circularity.NAME)
-public class DefaultCircularity<O extends RealType<O>> extends
-		AbstractGeometricFeature<Polygon, O> implements Geometric2D.Circularity {
+@Plugin(type = GeometricOp.class, label = "Geometric: Elongation", name = Geometric2D.Elongation.NAME)
+public class DefaultElongation<O extends RealType<O>> extends
+		AbstractGeometricFeature<Polygon, O> implements Geometric2D.Elongation
 
-	private FunctionOp<Polygon, O> areaFunc;
-	private FunctionOp<Polygon, O> perimiterFunc;
+{
+
+	private FunctionOp<Polygon, O> minorAxisFunc;
+	private FunctionOp<Polygon, O> majorAxisFunc;
 
 	@Override
 	public void initialize() {
-		areaFunc = RTs.function(ops(), Area.class, in());
-		perimiterFunc = RTs.function(ops(), Perimeter.class, in());
+		minorAxisFunc = RTs.function(ops(), MinorAxis.class, in());
+		majorAxisFunc = RTs.function(ops(), MajorAxis.class, in());
 	}
 
 	@Override
 	public void compute(final Polygon input, final O output) {
-		output.setReal(4
-				* Math.PI
-				* (areaFunc.compute(input).getRealDouble() / Math.pow(
-						perimiterFunc.compute(input).getRealDouble(), 2)));
+		output.setReal(1d - minorAxisFunc.compute(input).getRealDouble()
+				/ majorAxisFunc.compute(input).getRealDouble());
 	}
 
 }
