+++ src/main/java/net/imagej/ops/geom/DefaultFeret.java
@@ -27,41 +27,66 @@
  * POSSIBILITY OF SUCH DAMAGE.
  * #L%
  */
-package net.imagej.ops.geometric;
+package net.imagej.ops.geom;
 
-import net.imagej.ops.FunctionOp;
+import net.imagej.ops.AbstractFunctionOp;
+import net.imagej.ops.Contingent;
 import net.imagej.ops.Ops.Geometric2D;
-import net.imagej.ops.Ops.Geometric2D.Area;
-import net.imagej.ops.Ops.Geometric2D.MajorAxis;
-import net.imagej.ops.Ops.Geometric2D.Roundness;
-import net.imagej.ops.RTs;
+import net.imagej.ops.Ops.Geometric2D.Feret;
+import net.imglib2.RealLocalizable;
 import net.imglib2.roi.geometric.Polygon;
-import net.imglib2.type.numeric.RealType;
+import net.imglib2.util.Pair;
+import net.imglib2.util.ValuePair;
 
 import org.scijava.plugin.Plugin;
 
 /**
- * Generic implementation of {@link Roundness}.
+ * Generic implementation of {@link Feret}.
  * 
  * @author Daniel Seebacher, University of Konstanz.
  */
-@Plugin(type = GeometricOp.class, label = "Geometric: Roundness", name = Geometric2D.Roundness.NAME)
-public class DefaultRoundness<O extends RealType<O>> extends
-		AbstractGeometricFeature<Polygon, O> implements Geometric2D.Roundness {
-
-	private FunctionOp<Polygon, O> areaFunc;
-	private FunctionOp<Polygon, O> majorAxisFunc;
+@Plugin(type = GeometricOp.class, label = "Geometric: Feret", name = Geometric2D.Feret.NAME)
+public class DefaultFeret extends
+		AbstractFunctionOp<Polygon, Pair<RealLocalizable, RealLocalizable>>
+		implements
+		GeometricOp<Polygon, Pair<RealLocalizable, RealLocalizable>>,
+		Contingent, Geometric2D.Feret {
 
 	@Override
-	public void initialize() {
-		areaFunc = RTs.function(ops(), Area.class, in());
-		majorAxisFunc = RTs.function(ops(), MajorAxis.class, in());
+	public Pair<RealLocalizable, RealLocalizable> compute(final Polygon input) {
+
+		double distance = Double.NEGATIVE_INFINITY;
+		int in0 = -1;
+		int in1 = -1;
+
+		for (int i = 0; i < input.getVertices().size(); i++) {
+			for (int j = i + 1; j < input.getVertices().size(); j++) {
+				RealLocalizable temp0 = input.getVertices().get(i);
+				RealLocalizable temp1 = input.getVertices().get(j);
+
+				double sum = 0;
+				for (int k = 0; k < temp0.numDimensions(); k++) {
+					sum += Math.pow(
+							temp0.getDoublePosition(k)
+									- temp1.getDoublePosition(k), 2);
+				}
+				sum = Math.sqrt(sum);
+
+				if (sum > distance) {
+					distance = sum;
+					in0 = i;
+					in1 = j;
+				}
+			}
+		}
+
+		return new ValuePair<RealLocalizable, RealLocalizable>(input
+				.getVertices().get(in0), input.getVertices().get(in1));
 	}
 
 	@Override
-	public void compute(final Polygon input, final O output) {
-		output.setReal(4 * (areaFunc.compute(input).getRealDouble() / (Math.PI * Math
-				.pow(majorAxisFunc.compute(input).getRealDouble(), 2))));
+	public boolean conforms() {
+		return 2 == in().numDimensions();
 	}
 
 }
