+++ src/main/java/net/imagej/ops/geom/DefaultMinorAxis.java
@@ -27,53 +27,40 @@
  * POSSIBILITY OF SUCH DAMAGE.
  * #L%
  */
-package net.imagej.ops.geometric;
+package net.imagej.ops.geom;
 
-import net.imagej.ops.AbstractFunctionOp;
-import net.imagej.ops.Contingent;
+import net.imagej.ops.FunctionOp;
 import net.imagej.ops.Ops.Geometric2D;
-import net.imagej.ops.Ops.Geometric2D.Centroid;
-import net.imglib2.RealLocalizable;
-import net.imglib2.RealPoint;
+import net.imagej.ops.Ops.Geometric2D.MinorAxis;
+import net.imagej.ops.Ops.Geometric2D.MinorMajorAxis;
 import net.imglib2.roi.geometric.Polygon;
+import net.imglib2.type.numeric.RealType;
+import net.imglib2.util.Pair;
 
 import org.scijava.plugin.Plugin;
 
 /**
- * Generic implementation of {@link Centroid}.
+ * Generic implementation of {@link MinorAxis}.
  * 
  * @author Daniel Seebacher, University of Konstanz.
  */
-@Plugin(type = GeometricOp.class, label = "Geometric: Center of Gravity", name = Geometric2D.Centroid.NAME)
-public class DefaultCenterOfGravity extends
-		AbstractFunctionOp<Polygon, RealLocalizable> implements
-		GeometricOp<Polygon, RealLocalizable>, Contingent, Geometric2D.Centroid {
+@Plugin(type = GeometricOp.class, label = "Geometric: Minor Axis", name = Geometric2D.MinorAxis.NAME)
+public class DefaultMinorAxis<O extends RealType<O>> extends
+		AbstractGeometricFeature<Polygon, O> implements Geometric2D.MinorAxis {
+
+	@SuppressWarnings("rawtypes")
+	private FunctionOp<Polygon, Pair> minorMajorAxisFunc;
 
 	@Override
-	public RealLocalizable compute(final Polygon input) {
-
-		double area = 0;
-		double cx = 0;
-		double cy = 0;
-		for (int i = 0; i < input.getVertices().size() - 1; i++) {
-			RealLocalizable p0 = input.getVertices().get(i);
-			RealLocalizable p1 = input.getVertices().get(i + 1);
-
-			double p0_x = p0.getDoublePosition(0);
-			double p0_y = p0.getDoublePosition(1);
-			double p1_x = p1.getDoublePosition(0);
-			double p1_y = p1.getDoublePosition(1);
-
-			cx += (p0_x + p1_x) * (p0_x * p1_y - p1_x * p0_y);
-			cy += (p0_y + p1_y) * (p0_x * p1_y - p1_x * p0_y);
-		}
-
-		return new RealPoint(cx / (area * 6), cy / (area * 6));
+	public void initialize() {
+		minorMajorAxisFunc = ops().function(MinorMajorAxis.class, Pair.class,
+				in());
 	}
 
+	@SuppressWarnings("unchecked")
 	@Override
-	public boolean conforms() {
-		return 2 == in().numDimensions();
+	public void compute(final Polygon input, final O output) {
+		Pair<Double, Double> compute = minorMajorAxisFunc.compute(input);
+		output.setReal(compute.getA());
 	}
-
 }
