+++ src/main/java/net/imagej/ops/geom/RandomAccessibleIntervalToPolygonConverter.java
@@ -27,7 +27,7 @@
  * POSSIBILITY OF SUCH DAMAGE.
  * #L%
  */
-package net.imagej.ops.geometric;
+package net.imagej.ops.geom;
 
 import java.lang.reflect.Type;
